Made DecisionEngine cleaner; Renamed PipelineCreator and PrepLoaderTest
