+++ data-prep/dataprep-simple/src/main/scala/de/hpi/isg/dataprep/selection/GreedyPipelineCreator.scala
@@ -1,0 +1,20 @@
+package de.hpi.isg.dataprep.selection
+
+import de.hpi.isg.dataprep.components.{Pipeline, Preparation}
+import de.hpi.isg.dataprep.context.DataContext
+
+class GreedyPipelineCreator(dataContext: DataContext) extends PipelineCreator(dataContext) {
+  private val MAX_ITERATIONS = 100
+  private val engine = new DecisionEngine()
+
+  override def createPipeline(): Pipeline = {
+    val pipe = new Pipeline(dataContext)
+    for (_ <- 0 to MAX_ITERATIONS) {
+      engine.selectNextPreparation(preparators, dataContext.getSchemaMapping, dataContext.getDataFrame, dataContext.getTargetMetadata) match {
+        case Some(prep) => pipe.addPreparation(new Preparation(prep))
+        case None => return pipe
+      }
+    }
+    pipe
+  }
+}
