+++ data-prep/dataprep-simple/src/main/scala/de/hpi/isg/dataprep/selection/PipelineCreator.scala
@@ -1,10 +1,11 @@
 package de.hpi.isg.dataprep.selection
 
-import de.hpi.isg.dataprep.components.{Pipeline, Preparation}
+import de.hpi.isg.dataprep.components.Pipeline
 import de.hpi.isg.dataprep.context.DataContext
+import de.hpi.isg.dataprep.model.target.system.AbstractPreparator
 
 abstract class PipelineCreator(dataContext: DataContext) {
-  val preparators = PipelineCreator.preparators
+  val preparators: List[Class[_ <: AbstractPreparator]] = PipelineCreator.preparators
   def createPipeline(): Pipeline
 }
 
@@ -14,20 +15,3 @@ object PipelineCreator extends PreparatorLoader {
 }
 
 
-class SimplePipelineCreator(dataContext: DataContext) extends PipelineCreator(dataContext) {
-  private val MAX_ITERATIONS = 100
-  private val engine = new SimpleDecisionEngine()
-
-  override def createPipeline(): Pipeline = {
-    val pipe = new Pipeline(dataContext)
-
-
-    for (_ <- 0 to MAX_ITERATIONS) {
-      engine.selectNextPreparation(preparators, dataContext.getSchemaMapping, dataContext.getDataFrame, dataContext.getTargetMetadata) match {
-        case Some(prep) => pipe.addPreparation(new Preparation(prep))
-        case None => return pipe
-      }
-    }
-    pipe
-  }
-}
