+++ data-prep/dataprep-simple/src/test/scala/de/hpi/isg/dataprep/selection/PreparatorLoaderTest.scala
@@ -3,10 +3,9 @@ package de.hpi.isg.dataprep.selection
 import de.hpi.isg.dataprep.preparators.define.DeleteProperty
 import org.scalatest.{FlatSpecLike, Matchers}
 
-class ReflectionTest  extends FlatSpecLike with Matchers {
+class PreparatorLoaderTest  extends FlatSpecLike with Matchers {
 
-
-  it should "load all preparator" in {
+  "The PreparatorLoader" should "load a preparator" in {
     class TestLoader extends PreparatorLoader {
       override val path: String =  "de.hpi.isg.preparators"
     }
